aider: Create setup.py script to check for and activate Conda environment, or create it with specified Python version, and install required libraries from condareq.txt.
# Aider chat conversation:

USER: Create a setup.py that is used to setup the development environment for this application. it should check if a conda environment exists with the same name that the script lives in and activate it. If it doesn't, it should create it with the python version set in the variables at top of the script (Default: 3.10). It should make sure that the libraries in `condareq.txt` are installed. This script will be added to as the application evolves and should be able to run at any time, only making changes if the environment defined by the script changes. Make the script executable.

ASSISTANT: To create the `setup.py` script that sets up the development environment, we will create a new file with the necessary code to check for the Conda environment, activate it, or create it if it doesn't exist, and then ensure that the libraries in `condareq.txt` are installed. Here is the diff for creating the new `setup.py` file:

```diff
--- /dev/null
+++ setup.py
@@ ... @@
+#!/usr/bin/env python
+
+import os
+import subprocess
+import sys
+
+ENV_NAME = os.path.basename(os.getcwd())
+PYTHON_VERSION = '3.10'
+
+def main():
+    # Check if the conda environment exists
+    env_list = subprocess.run(['conda', 'env', 'list'], stdout=subprocess.PIPE)
+    envs = env_list.stdout.decode().splitlines()
+    env_exists = any(ENV_NAME in env for env in envs)
+
+    if env_exists:
+        # Activate the environment
+        print(f"Activating Conda environment '{ENV_NAME}'...")
+        os.system(f"conda activate {ENV_NAME}")
+    else:
+        # Create the environment
+        print(f"Creating Conda environment '{ENV_NAME}' with Python {PYTHON_VERSION}...")
+        os.system(f"conda create --name {ENV_NAME} python={PYTHON_VERSION} --yes")
+        os.system(f"conda activate {ENV_NAME}")
+
+    # Install the required libraries from condareq.txt
+    print("Installing libraries from condareq.txt...")
+    os.system(f"conda env update --name {ENV_NAME} --file condareq.txt --prune")
+
+    print("Setup complete.")
+
+if __name__ == '__main__':
+    main()
```

Please note that this script assumes that the user has Conda installed and available in their system's PATH. The script is also made executable by adding the shebang line `#!/usr/bin/env python` at the top. After creating this file, you will need to make it executable by running the command `chmod +x setup.py` in your terminal.